+++ pkg/runners/namespaces.go
@@ -1,5 +1,5 @@
 // pkg/reconciler/run_namespace.go
-package reconciler
+package runners
 
 import (
 	"context"
@@ -13,44 +13,7 @@ import (
 	orktypes "github.com/orkspace/orkestra/pkg/types"
 )
 
-// deleteOwnedNamespaces explicitly deletes all Namespaces declared across
-// onCreate and onReconcile that are owned by this CR.
-//
-// Kubernetes GC does not handle this automatically: owner references from
-// namespace-scoped resources (CRs) to cluster-scoped resources (Namespaces)
-// are not honoured by the garbage collector. Explicit deletion is required.
-func deleteOwnedNamespaces(
-	ctx context.Context,
-	kube kubeclient.KubeClient,
-	resolver *orktmpl.Resolver,
-	obj domain.Object,
-	box orktypes.OperatorBoxConfig,
-) error {
-	// Collect all declared namespace sources across template blocks.
-	var srcs []orktypes.NamespaceTemplateSource
-	if box.OnCreate != nil {
-		srcs = append(srcs, box.OnCreate.Namespaces...)
-	}
-	if box.OnReconcile != nil {
-		srcs = append(srcs, box.OnReconcile.Namespaces...)
-	}
-	if box.OnDelete != nil {
-		srcs = append(srcs, box.OnDelete.Namespaces...)
-	}
-
-	for i, src := range srcs {
-		name, err := resolver.Resolve(src.Name)
-		if err != nil || name == "" {
-			continue
-		}
-		if err := orkns.DeleteIfOwned(ctx, kube, obj, name); err != nil {
-			return fmt.Errorf("namespace[%d] %q: %w", i, name, err)
-		}
-	}
-	return nil
-}
-
-// runNamespaces resolves and applies Namespace template declarations.
+// RunNamespaces resolves and applies Namespace template declarations.
 //
 // Namespaces are create-only — there is nothing meaningful to update
 // on a Namespace after creation (no spec fields that drift). They are
@@ -58,7 +21,7 @@ func deleteOwnedNamespaces(
 // from onCreate or onReconcile.
 //
 // Owner references ensure cleanup when the CR is deleted.
-func runNamespaces(
+func RunNamespaces(
 	ctx context.Context,
 	kube kubeclient.KubeClient,
 	resolver *orktmpl.Resolver,
